refactor: 폴더명 수정,시멘틱 태그 사용 및 중복확인 버튼 위치 버그 수정
refactor: 컨벤션을 위해 폴더명 수정

refactor: 중복확인 버튼 고정된 위치 버그 수정

refactor: 시멘틱 태그 label 사용
