+++ src/api/index.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { OWHAT_TOKEN } from '~/constant/token';
+import { OWHAT_TOKEN } from '~/constants/token';
 import { BrowserStorage } from '~/utils/storage';
 
 const BASE_URL = import.meta.env.VITE_BASE_URL;
